src/test/java/com/example/controllers/AdminControllerTest.java
package com.example.controllers;

import com.example.entities.User;
import com.example.exeption.MyException;
import com.example.models.common.UserDto;
import com.example.models.request.CreateTaskRequestDto;
import com.example.services.UserService;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.dao.EmptyResultDataAccessException;
import org.springframework.http.MediaType;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import static org.junit.jupiter.api.Assertions.*;

@RunWith(SpringRunner.class)
@AutoConfigureMockMvc(addFilters = false)
@WebMvcTest(AdminController.class)
class AdminControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @Autowired
    private AdminController adminController;

    @Autowired
    private ObjectMapper objectMapper;

    @MockBean
    private UserService userService;

    @Test
    void createUser() throws Exception {

        UserDto userDto = new UserDto("Admin", "Admin");
        User user = new User();
        user.setId(1l);
        user.setUserName("Admin");
        user.setPassword("Admin");
        Mockito.when(userService.saveUser(userDto)).thenReturn(true);
        Mockito.when(userService.loadUserByUsername(userDto.getUserName())).thenReturn(user);

        mockMvc.perform(MockMvcRequestBuilders.post("/admin")
                .content(objectMapper.writeValueAsString(userDto))
                .contentType(MediaType.APPLICATION_JSON))

                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.content().contentType(MediaType.APPLICATION_JSON));

    }

    @Test
    void deleteTask() throws Exception {

        mockMvc.perform(MockMvcRequestBuilders.delete("/admin/1"))

                .andExpect(MockMvcResultMatchers.status().isOk());

    }
}